Conflicts:
	MPC.Web/Areas/Api/Controllers/CostCenterTreeController.cs
	MPC.Web/MPC.MIS.csproj

+++ Repository/Repositories/ItemRepository.cs
@@ -195,18 +195,18 @@ namespace MPC.Repository.Repositories
         {
             try
             {
-                if (CompanyId > 0)
-                {
+                //if (CompanyId > 0)
+                //{
+                //    return
+                //        db.ItemStockOptions.Where(
+                //            i => i.ItemId == ItemId && i.CompanyId == CompanyId && i.OptionSequence == 1).FirstOrDefault();
+                //}
+                //else
+                //{
                     return
-                        db.ItemStockOptions.Where(
-                            i => i.ItemId == ItemId && i.CompanyId == CompanyId && i.OptionSequence == 1).FirstOrDefault();
-                }
-                else
-                {
-                    return
-                        db.ItemStockOptions.Where(i => i.ItemId == ItemId && i.CompanyId == CompanyId && i.OptionSequence == 1)
+                        db.ItemStockOptions.Where(i => i.ItemId == ItemId && i.OptionSequence == 1)
                             .FirstOrDefault();
-                }
+                //}
             }
             catch (Exception ex)
             {
@@ -594,16 +594,19 @@ namespace MPC.Repository.Repositories
 
                             }
                         }
-
-                        File.Copy(drURL + oTemplatePage.BackgroundFileName,
-                            drURL + result.ToString() + "/" +
-                            oTemplatePage.BackgroundFileName.Substring(oTemplatePage.BackgroundFileName.IndexOf("/"),
-                                oTemplatePage.BackgroundFileName.Length - oTemplatePage.BackgroundFileName.IndexOf("/")));
-                        oTemplatePage.BackgroundFileName = result.ToString() + "/" +
-                                                           oTemplatePage.BackgroundFileName.Substring(
-                                                               oTemplatePage.BackgroundFileName.IndexOf("/"),
-                                                               oTemplatePage.BackgroundFileName.Length -
-                                                               oTemplatePage.BackgroundFileName.IndexOf("/"));
+                        if (File.Exists(drURL + oTemplatePage.BackgroundFileName))
+                        {
+                            File.Copy(drURL + oTemplatePage.BackgroundFileName,
+                                drURL + result.ToString() + "/" +
+                                oTemplatePage.BackgroundFileName.Substring(oTemplatePage.BackgroundFileName.IndexOf("/"),
+                                    oTemplatePage.BackgroundFileName.Length - oTemplatePage.BackgroundFileName.IndexOf("/")));
+                            oTemplatePage.BackgroundFileName = result.ToString() + "/" +
+                                                               oTemplatePage.BackgroundFileName.Substring(
+                                                                   oTemplatePage.BackgroundFileName.IndexOf("/"),
+                                                                   oTemplatePage.BackgroundFileName.Length -
+                                                                   oTemplatePage.BackgroundFileName.IndexOf("/"));
+                        }
+                        
 
                     }
                 }
